requirements.txt
# MCP SDK для создания сервера
mcp[cli]>=1.0.0

# HTTP клиент для запросов к API
httpx>=0.27.0

# Работа с переменными окружения
python-dotenv>=1.0.0

# OpenAI API для клиента
openai>=1.12.0

# MCP клиент
mcp>=1.0.0
